adjustments to email results. note if email not set up, error at present.

+++ Drive/SelectCarViewController.m
@@ -47,7 +47,7 @@
     //levelVC.carType = sender.tag;
     
     //singleton.carNo=[NSString stringWithFormat:@"%ld", levelVC.carType];
-    singleton.carNo=[NSString stringWithFormat:@"%ld", sender.tag];
+    singleton.carNo=[NSString stringWithFormat:@"%ld", (long)sender.tag];
 
     //[self.navigationController pushViewController:levelVC animated:YES];
 }
